+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
+- nothing yet
+
+## [v1.5.0]
 ### Added
 - Optionally retain s3 blobs on message delete [#16](https://github.com/PruvoNet/squiss-ts/issues/16)
 - Optionally set prefix for s3 blob names [#15](https://github.com/PruvoNet/squiss-ts/issues/15)
@@ -73,3 +76,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [v1.2.4]: https://github.com/PruvoNet/squiss-ts/compare/v1.2.3...v1.2.4
 [v1.3.0]: https://github.com/PruvoNet/squiss-ts/compare/v1.2.4...v1.3.0
 [v1.4.0]: https://github.com/PruvoNet/squiss-ts/compare/v1.3.0...v1.4.0
+[v1.5.0]: https://github.com/PruvoNet/squiss-ts/compare/v1.4.0...v1.5.0